pyproject.toml
[project]
name = "uv-pg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.9.2",
    "pyright>=1.1.384",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.9",
    "typer>=0.12.5",
]
